requirements.txt
beautifulsoup4==4.12.3
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
pandas==2.2.3
plotly==5.24.1
httplib2==0.22.0
matplotlib==3.9.2
oauthlib==3.2.2
ply==3.11
proto-plus==1.25.0
protobuf==4.25.3
selenium==4.24.0
streamlit==1.41.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyQt5==5.15.10
pywin32==307
requests-oauthlib==2.0.0
rsa==4.9
streamlit-card==1.0.2
streamlit-tags==1.2.8
streamlit-card==1.0.2
streamlit-tags==1.2.8
undetected-chromedriver==3.5.5
uritemplate==4.1.1
websockets==14.2
zstandard==0.23.0
